Update Test-IfModulesInBlobStorage.ps1
Hello!

I am currently implementing the M365DSC project in our company. I have noticed that in the Test-IfModulesInBlobStorage Function, the if part at the end does not make sense. Whether true or not, you always get an error that the zip file was not found in the BlobStorage. Here is my suggested change. Please give me feedback. Best regards, Erik